Introduce operations on Instances, implements #464 (#602)
This commit adds initial support for operations on the instances table. At present sync adapters are not allowed to write to the instances table, non-sync adapters can only insert or modify task values. Instances table columns are still read-only. Also, columns related to recurrence can not be set via the instances table.
This is meant to simplify working with recurrence in the UI. The UI won't have to take special care about creating exceptions. It can simply update or delete individual instances and the provider will take care of creating overrides or ex-dates.

This also fixes Travis not running the `check` task and improves/fixes the provide tests in general.