Let instance view return instance DTSTART and DUE rather than the task values. Implements #615
This commit improves the instances view by returning the instance start and due times as task `DTSTART` and `DUE`. This way the UI will see the correct instance times.

From the UI perspective the instances view is now a list of individual non-recurring instances, which can be read and written on its own.

Note, for non-recurring tasks, this won't have any effect.