Fix wrong ID being used. Fixes #723 (#724)
In an earlier commit we changed how the lists are loaded. Instead of loading them from the tasks table we now load them from the instances table. This was meant to prepare support for recurring tasks.

Unfortunately we missed on line where we were still using the `_id` column to get the task id (when we should actually use the `task_id` column now). This didn't cause any problem as long as the task id and the instance id where in sync (hence it wasn't an issue while testing). However, under certain circumstances this assumption is no longer true and caused an illegal Task URI to be loaded when clicking a task.

The list now uses the `task_id` column to create the task URI to show.