Squashed commit of the discussed modifications
Removed typos

Added title to notification relevant fields

Reason is if a sync process updates the name of a task, a notification
might already be showing. In this case the notification text should be
updated according to the new title.
The same holds true for all shown information in the notification. As the
exact graphical representation of the notification is not fixed, this is
thus not 100% fixed as well.