Create version column if required, fixes #793 (#794)
The cause of issue #793 is actually unknown but this once could be a good candidate. In few cases the database might already be on version 20 but the version column had not been created.
This might be caused by having the beta version installed before (which already had a version 20, but without the column).

This fix always checks if the version column already exists and creates it otherwise.