fix creating deleted instances on move, fixes #829
Upon moving a task to a new list, it's not just moved, but a deleted copy is left behind in the old list, so sync adapters can take note of the removal in the old list.
This commit fixes the creation of the deleted instance. The code tried to create instance columns in the task table, which failed and caused the delted instance to be missing.

This commit ensures that the TaskAdapter which is used to insert the deleted instance does not contain instance values.